requirements.txt
aiogram==3.2.0
django-environ==0.11.2
peewee==3.17.0
